# docs/openapi/facilities.yaml
openapi: 3.1.0
info:
  title: SolarSimz 주요설비 API
  version: 1.0.0
  description: 대산금속 주요설비 등록, 조회, 수정, 삭제 및 복구 API
servers:
  - url: http://localhost:3456
security:
  - sessionCookie: []
paths:
  /api/tokens:
    post:
      summary: 로그인 및 HttpOnly 세션 발급
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 원본 대시보드 호환 로그인 응답과 세션 쿠키
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          description: 로그인 시도 제한
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: solar_session
...
  schemas:
    LoginRequest:
      type: object
      required: [cf, id, pw]
      additionalProperties: false
      properties:
        cf: { const: login }
        id: { type: string, minLength: 1, maxLength: 80 }
        pw: { type: string, minLength: 1, maxLength: 256 }
    LoginResponse:
      type: object
      required: [token, authIdn, authName, fid, firmName, role, members]
      properties:
        token: { type: string }
        authIdn: { type: string, format: uuid }
        authName: { type: string }
        fid: { type: [string, integer] }
        firmName: { type: string }
        role: { type: string, enum: [ADMIN, OPERATOR, VIEWER] }
        members:
          type: array
          description: 현재 테넌트에 속한 업체 선택 항목만 반환
          items:
            type: object
            additionalProperties: true
...
    Error:
      type: object
      required: [ok, requestId, error]
      properties:
        ok: { const: false }
        requestId: { type: string, format: uuid }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fieldErrors:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
...
  responses:
    Unauthorized:
      description: 로그인 필요
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: 입력 검증 실패
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }